+++ packages/big-design/src/components/StatefulTable/StatefulTable.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useEffect, useMemo, useReducer } from 'react';
 import { useDidUpdate } from '../../hooks';
 import { typedMemo } from '../../utils';
 import { Box } from '../Box';
-import { PaginationProps } from '../Pagination';
-import { OffsetPaginationLocalization } from '../Pagination/Pagination';
+import { PaginationProps } from '../OffsetPagination';
+import { OffsetPaginationLocalization } from '../OffsetPagination/OffsetPagination';
 import { PillTabItem, PillTabs, PillTabsProps } from '../PillTabs';
 import { Search } from '../Search';
 import { SearchLocalization } from '../Search/types';
